add error boundary for send route
