+++ deployables/app/app/routes/tokens/send/layout.tsx
@@ -1,0 +1,14 @@
+import { Page } from '@/components'
+import { Outlet } from 'react-router'
+
+const SendLayout = () => (
+  <Page>
+    <Page.Header>Send tokens</Page.Header>
+
+    <Page.Main>
+      <Outlet />
+    </Page.Main>
+  </Page>
+)
+
+export default SendLayout
